# Fix user and kernel times on Windows