--- a/Cargo.toml
+++ b/Cargo.toml
@@ -11,6 +11,10 @@ version = "1.14.0"
 edition = "2018"
 build = "build.rs"
 
+[features]
+# Use the nightly feature windows_process_extensions_main_thread_handle
+windows_process_extensions_main_thread_handle = []
+
 [dependencies]
 colored = "2.0"
 indicatif = "0.16"
@@ -29,7 +33,8 @@ anyhow = "1.0"
 libc = "0.2"
 
 [target.'cfg(windows)'.dependencies]
-winapi = { version = "0.3", features = ["processthreadsapi", "minwindef", "winnt"] }
+winapi = { version = "0.3", features = ["processthreadsapi", "minwindef", "winnt", "jobapi2", "tlhelp32"] }
+once_cell = "1.13"
 
 [target.'cfg(target_os="linux")'.dependencies]
 nix = { version = "0.24.2", features = ["zerocopy"] }