@clemenswasser 2022-08-15
@sharkdp thanks for your comments. I'll look into them tomorrow. I already found a undocumented ntdll function NtResumeProcess, which resumes a process by HANDLE, which we could potentially use and before that I played around with a implementation using CreateToolhelp32Snapshot for finding the main thread, but using NtResumeProcess should definitely be better in terms of performance and code size 😁.