@clemenswasser 2022-08-16
With the latest commit it works on stable using the undocumented but well known `ntdll.dll!NtResumeThread` which exists since Windows XP. When using nightly the feature `windows_process_extensions_main_thread_handle` can be set to use the unstable feature with the same name. The dependency on once_init is not a problem, since clap was already using it anyway.
Also I don't know why the `ActiveProcesses == 0` assertion triggered on https://github.com/sharkdp/hyperfine/runs/7864408316...
For me locally this never triggered for simple test cases like `cargo run -- 'cargo fmt'`